requirements.txt
discord.py
deep-translator
langdetect
flask
pillow
pytesseract
python-dotenv
requests